% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFlower.R
\name{venn.flower}
\alias{venn.flower}
\title{Flower venn plot}
\usage{
venn.flower(
  df,
  start = 90,
  a = 0.5,
  b = 2,
  r = 1,
  ellipse_col = c("#6181BD4E", "#F348004E", "#64A10E4E", "#9300264E", "#464E044E",
    "#049a0b4E", "#4E0C664E", "#D000004E", "#FF6C004E", "#FF00FF4E", "#c7475b4E",
    "#00F5FF4E", "#BDA5004E", "#A5CFED4E", "#f0301c4E", "#2B8BC34E", "#FDA1004E",
    "#54adf54E", "#CDD7E24E", "#9295C14E", "#6181BD4E", "#FDA1004E"),
  circle_col = "white",
  ...
)
}
\arguments{
\item{df}{A data frame, each column typically represents a different sample or group, and the rows represent the observations or features.}

\item{start}{A numeric value that specifies the starting angle (in degrees) for the first ellipse in the flower plot (default: 90).}

\item{a}{A numeric value representing the length of the semi-major axis of the ellipses (default: 0.5).}

\item{b}{A numeric value representing the length of the semi-minor axis of the ellipses (default: 2).}

\item{r}{A numeric value that specifies the radius of the central circle in the flower plot (default: 1).}

\item{ellipse_col}{A vector of colors for the ellipses.}

\item{circle_col}{A character string specifying the color of the central circle in the plot.}

\item{...}{Additional graphical parameters that can be passed to the plotting functions. This allows for further customization of the plot, such as adjusting text size, font, or other graphical parameters.}
}
\value{
Venn flower plot
}
\description{
Flower venn plot
}
\examples{
library(vennFlower)
library(dplyr)
library(plotrix)
my_col <- c(
  "#6181BD4E", "#F348004E", "#64A10E4E", "#9300264E",
  "#464E044E", "#049a0b4E", "#4E0C664E", "#D000004E",
  "#FF6C004E", "#FF00FF4E", "#c7475b4E", "#00F5FF4E",
  "#BDA5004E", "#A5CFED4E", "#f0301c4E", "#2B8BC34E",
  "#FDA1004E", "#54adf54E", "#CDD7E24E", "#9295C14E",
  "#6181BD4E", "#FDA1004E")
data("gene_absence")
p <- venn.flower(df = gene_absence, ellipse_col = my_col)
# If you want to save the venn flower results, use:
png("venn_flower.png", width = 10, height = 10, units = "in", res = 300)
p
dev.off()
}
